🔄 synced local 'skyvern-frontend/src/' with remote 'skyvern-frontend/src/'
Part of improving debugger.

- componentize window component, so it can be reused: `FloatingWindow`
- stub in min/max/restore/close buttons
  - both macOS and Windows style
  - only enable max/restore

<img width="3420" height="2214" alt="image" src="https://github.com/user-attachments/assets/c9e58447-a91e-4116-9f85-4522471d0017" />

<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Add reusable `FloatingWindow` component with maximize/restore buttons and integrate into workflow debug overview.
>
>   - **Components**:
>     - Add `FloatingWindow` component in `FloatingWindow.tsx` for draggable, resizable windows.
>     - Integrate `FloatingWindow` into `WorkflowDebugOverviewWindow.tsx`.
>   - **Buttons**:
>     - Add stub buttons for min/max/restore/close in `FloatingWindow.tsx`.
>     - Enable maximize and restore functionality for both macOS and Windows styles.
>
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=Skyvern-AI%2Fskyvern-cloud&utm_source=github&utm_medium=referral)<sup> for 0112f2cc8b5efe6025e142900c15efc4a93a098b. You can [customize](https://app.ellipsis.dev/Skyvern-AI/settings/summaries) this summary. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->
